# apps/api/security/openapi.json
{
  "components": {
    "schemas": {
...
      "TemplateEnvelope": {
        "properties": {
          "template": {
            "$ref": "#/components/schemas/TemplateSchema"
          }
        },
        "required": [
          "template"
        ],
        "type": "object"
      },
      "TemplateListEnvelope": {
        "properties": {
          "templates": {
            "items": {
              "$ref": "#/components/schemas/TemplateSchema"
            },
            "type": "array"
          }
        },
        "required": [
          "templates"
        ],
        "type": "object"
      },
      "TemplateSchema": {
        "properties": {
          "container_runtime": {
            "type": "string"
          },
          "cores": {
            "format": "int32",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "docker_in_instance": {
            "type": "boolean"
          },
          "docker_registry": {
            "type": [
              "string",
              "null"
            ]
          },
          "exec_config": {},
          "gpu_count": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "instance_count": {
            "format": "int64",
            "type": "integer"
          },
          "keep_time_action": {
            "type": "string"
          },
          "keep_time_seconds": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "max_run_seconds": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "memory": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "network_bandwidth_down_mbps": {
            "format": "int32",
            "type": "integer"
          },
          "network_bandwidth_up_mbps": {
            "format": "int32",
            "type": "integer"
          },
          "owner_id": {
            "format": "uuid",
            "type": "string"
          },
          "persistent_storage_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "remote_type": {
            "type": "string"
          },
          "run_config": {},
          "timeout_action": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "visibility": {
            "type": "string"
          },
          "volume_mappings": {}
        },
        "required": [
          "id",
          "name",
          "owner_id",
          "image",
          "cores",
          "memory",
          "gpu_count",
          "remote_type",
          "container_runtime",
          "run_config",
          "exec_config",
          "volume_mappings",
          "timeout_action",
          "keep_time_action",
          "network_bandwidth_up_mbps",
          "network_bandwidth_down_mbps",
          "docker_in_instance",
          "visibility",
          "instance_count",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "OpenAPI spec for the 17 security-fuzzable endpoints (see .scratch/security-fuzzing). The 8 admin-gated operations are tagged `admin-gated` for the low-privilege RBAC-boundary pass. The spec is export-only and never served.",
    "license": {
      "name": ""
    },
    "title": "OpenWorkspace API — security fuzz surface",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/templates": {
      "get": {
        "operationId": "list_templates",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateListEnvelope"
                }
              }
            },
            "description": "templates catalog"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "templates"
        ]
      }
    },
    "/api/templates/{id}": {
      "get": {
        "operationId": "get_template",
        "parameters": [
          {
            "description": "template uuid",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateEnvelope"
                }
              }
            },
            "description": "template detail"
          },
          "400": {
            "description": "invalid uuid"
          },
          "401": {
            "description": "missing or invalid ow_token"
          },
          "404": {
            "description": "template not found"
          },
          "500": {
            "description": "internal server error"
          }
        },
        "tags": [
          "templates"
        ]
      }
    },
...
  }
}